third_party/WebKit/LayoutTests/http/tests/media/media-source/stream_memory_tests/mediasource-appendstream-quota-exceeded-default-buffers.html

<!DOCTYPE html>
<html>
    <head>
        <script src="/w3c/resources/testharness.js"></script>
        <script src="/w3c/resources/testharnessreport.js"></script>
        <script src="../mediasource-util.js"></script>
    </head>
    <body>
        <script>
          var mediaURL = "/media/resources/media-source/webm/test-a-5min-44100Hz-1ch.webm";

          mediasource_test(function(test, mediaElement, mediaSource)
          {
              var sourceBuffer = mediaSource.addSourceBuffer(MediaSourceUtil.AUDIO_ONLY_TYPE);
              sourceBuffer.mode = 'sequence';

              MediaSourceUtil.fillUpSourceBufferViaAppendStream(test, mediaSource, sourceBuffer, mediaURL, function (appendedDataSize) {
                  assert_greater_than(appendedDataSize, 12 * 1048576, "Appended more than 12MB of data");
                  assert_less_than(appendedDataSize, 14 * 1048576, "Appended less than 14MB of data");
                  test.done();
              }, 551 * 1024);
          }, 'Calling appendStream repeatedly with size parameter should fill up the buffer and throw a QuotaExceededError when buffer is full.');

          mediasource_test(function(test, mediaElement, mediaSource)
          {
              var sourceBuffer = mediaSource.addSourceBuffer(MediaSourceUtil.AUDIO_ONLY_TYPE);
              sourceBuffer.mode = 'sequence';

              // When SourceBuffer.appendStream is called without the explicit size parameter, then the buffer full condition is reported
              // by posting an 'error' event on the SourceBuffer. There should be no QuotaExceededError exception in this case.
              sourceBuffer.addEventListener('error',  test.step_func_done());

              MediaSourceUtil.fillUpSourceBufferViaAppendStream(test, mediaSource, sourceBuffer, mediaURL, function (appendedDataSize) {
                  assert_unreached('Unexpected QuotaExceededError exception');
              });
          }, 'Calling appendStream repeatedly without size parameter should fill up the buffer and send an error event when buffer is full.');
        </script>
    </body>
</html>
